add native method deploying serverless app
