+++ examples/ecs-apps-monorepo/apps/pecan/.gitignore
@@ -1,0 +1,4 @@
+.serverless
+*.pyc
+*.pyo
+*node_modules*
